v7.0.0-alpha.1
Added the 5 basic classes that can be used to represent the different
stages of graph building. There are still many parts that need to be
implemented to reproduce what the scripts do. Also, the idea is to, in
the end, add a method to merge graphs by taking advantage the Graph
class. It will likely be static method for that class.