Create Entity class for OpenIE
This class is responsible for keeping entity references. This includes
keeping the tokens, the text, the dependency graph and the entity
mentions found in the text.

It is also responsible for building subtriples when the entity is made
up of other entities identified in the text. The ideia is to extract the
main entity from the pool of subentities and creating links between the
main entity and the secondary ones.