[OpenIE] Force subset entities to be smaller than main entities
There was an issue with entities that contained the same tokens in
different orders being identified as subsets of each other, resulting in
endless recursion. Just checking sizes was enough for the dataset being
used, but, in the future, it might be worth it to build sets out of the
tokens and check that the sets are not equivalent.