Add build_kg_bridges.py
This will replace doc_similarity.py and group_graph.py. The idea is to
treat each document as a different context and to use word embeddings to
discover links between contexts. With this approach, it is no longer
necessary to build one gigantic graph containing all entities. All
graphs are compared 2 by 2.