Ensure only one bridge match per entity
By just checking the maximum value for each row (axis=1), it was
possible to have two entities with the same bridge if both had their max
values at the same column. To fix this and ensure that each entity only
has one bridge, it is necessary to cross the maximum of the row with the
maximum of the column and just keep the maximums that appear in both.