Create runner for the KG generation
This runner will execute the four stages in sequence, skipping those
that are complete or that had their previous stage fail. That means that
it can also be used to resume processing in case it was interrupted for
some reason.

It is also worth poiting out that the idea is to run in the background,
in a separate process.