Have the graph builder behave like other builders
It now lets the user select one batch/group instead of letting them
select files. This is already how the other three builders work, so this
change should make the behavior of the builders more coherent.

This change also allows the public build_graphs function to be used,
which ensures the metadata will be updated.