Have runners save hyperparameters
Instead of the builder functions saving the parameters, the runners that
call the builder functions will now have that responsibility. The idea
behind this is that, considering the next step is to add pause/resume
features, we do not want the builders overwriting the parameters when
they are resumed, even if they would overwrite them with the exact same
values. The point is that it should not be their responsibility.