Add method to pause interrupted processes
If the server shuts down while processing a batch in parallel, that
process is interrupted. Because of that, it is necessary to have a way
to identify that that process needs to be resumed once the server is
restarted. The simplest way to do that is to check, on startup, which
batch stages are marked as "started", so they can be changed to
"paused". This only works at startup because a stage will only have the
"started" status at startup if it was interrupted.